+++ src/components/molecules/forms/fieldset.module.scss
@@ -5,12 +5,12 @@
   font-weight: 600;
 
   &#{&}--has-tooltip {
-    padding: 0 var(--spacing-xs) 0 0;
+    margin: 0 var(--spacing-2xs) 0 0;
   }
 }
 
 .btn {
-  margin: 0 var(--spacing-2xs) var(--spacing-2xs) 0;
+  margin: 0 var(--spacing-xs) var(--spacing-2xs) 0;
 
   &--activated {
     background: var(--color-primary);
