+++ src/components/molecules/forms/radio-group.stories.tsx
@@ -26,6 +26,19 @@ export default {
         required: false,
       },
     },
+    buttonClassName: {
+      control: {
+        type: 'text',
+      },
+      description: 'Set additional classnames to the help button.',
+      table: {
+        category: 'Styles',
+      },
+      type: {
+        name: 'string',
+        required: false,
+      },
+    },
     className: {
       control: {
         type: 'text',
