--- src/components/organisms/forms/settings-form.module.scss
@@ -1,81 +1,0 @@
-@use "@styles/abstracts/mixins" as mix;
-
-.wrapper {
-  display: flex;
-  flex-flow: row wrap;
-  align-items: flex-start;
-  align-content: flex-start;
-
-  @include mix.media("screen") {
-    @include mix.dimensions(null, "2xs", "height") {
-      column-gap: var(--spacing-lg);
-      row-gap: var(--spacing-xs);
-      font-size: var(--font-size-sm);
-    }
-  }
-
-  .label {
-    @include mix.media("screen") {
-      @include mix.dimensions(null, "2xs", "height") {
-        margin: 0 auto;
-        float: none;
-        font-size: var(--font-size-sm);
-      }
-
-      @include mix.dimensions(null, "2xs") {
-        //font-size: var(--font-size-sm);
-      }
-    }
-  }
-
-  .tooltip {
-    top: unset;
-    bottom: calc(100% + var(--spacing-2xs));
-    font-size: var(--font-size-sm);
-    transform-origin: bottom center;
-
-    @include mix.media("screen") {
-      @include mix.dimensions(null, "2xs", "height") {
-        width: 250%;
-        transform-origin: bottom left;
-      }
-
-      @include mix.dimensions("sm") {
-        font-size: var(--font-size-md);
-      }
-    }
-  }
-}
-
-.items {
-  margin: var(--spacing-2xs) 0;
-
-  @include mix.media("screen") {
-    @include mix.dimensions(null, "2xs", "height") {
-      display: flex;
-      flex-flow: column wrap;
-      max-width: fit-content;
-      margin: 0;
-    }
-  }
-}
-
-.group {
-  margin-left: auto;
-
-  @include mix.media("screen") {
-    @include mix.dimensions(null, "2xs", "height") {
-      margin: auto;
-    }
-  }
-}
-
-.fieldset__body {
-  margin-left: auto;
-
-  @include mix.media("screen") {
-    @include mix.dimensions(null, "2xs", "height") {
-      margin: 0 auto;
-    }
-  }
-}
