+++ src/components/organisms/modals/settings-modal.test.tsx
@@ -1,14 +1,40 @@
+import { storageKey as ackeeStorageKey } from '@components/molecules/forms/ackee-toggle.fixture';
+import { storageKey as motionStorageKey } from '@components/molecules/forms/motion-toggle.fixture';
 import { render, screen } from '@tests/utils';
 import SettingsModal from './settings-modal';
 
 describe('SettingsModal', () => {
-  it('renders a fake heading', () => {
+  it('renders the modal heading', () => {
     render(
       <SettingsModal
-        ackeeStorageKey="ackee-tracking"
-        motionStorageKey="reduce-motion"
+        ackeeStorageKey={ackeeStorageKey}
+        motionStorageKey={motionStorageKey}
       />
     );
     expect(screen.getByText(/Settings/i)).toBeInTheDocument();
+  });
+
+  it('renders a settings form', () => {
+    render(
+      <SettingsModal
+        ackeeStorageKey={ackeeStorageKey}
+        motionStorageKey={motionStorageKey}
+      />
+    );
+    expect(
+      screen.getByRole('form', { name: /^Settings form/i })
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('radiogroup', { name: /^Theme:/i })
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('radiogroup', { name: /^Code blocks:/i })
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('radiogroup', { name: /^Animations:/i })
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('radiogroup', { name: /^Tracking:/i })
+    ).toBeInTheDocument();
   });
 });
